+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### 🐞 Fixes
 
 - [#660](https://github.com/estruyf/vscode-front-matter/issues/660): Allow only to select unique content relationship values
+- [#661](https://github.com/estruyf/vscode-front-matter/issues/661): Fixing the dropdowns when used at the bottom of a collapsible group
 
 ## [9.2.0] - 2023-09-11